Address errors introduced by the latest version of silicon & Simplify code (#131)
With the new version of silicon, verification time has increased significantly. This package used to take 2min40s before and now seems to take around 5min to verify. This is likely aggravated by the removal of some outline blocks. I would argue that this is justifiable given how simpler the code becomes.

I only simplified part of the methods. I believe that we can achieve similar simplifications by redoing the remaining methods but I will not do it atm for the sake of time.